qgm: split `TestCatalog` in a separte submodule
