+++ src/sql/src/query_model/test/catalog.rs
@@ -14,13 +14,11 @@ use crate::catalog::{
 };
 use crate::func::{Func, MZ_CATALOG_BUILTINS, MZ_INTERNAL_BUILTINS, PG_CATALOG_BUILTINS};
 use crate::names::{DatabaseSpecifier, FullName, PartialName};
-use crate::plan::query::QueryLifetime;
-use crate::plan::{StatementContext, StatementDesc};
+use crate::plan::StatementDesc;
 use build_info::DUMMY_BUILD_INFO;
 use chrono::MIN_DATETIME;
 use dataflow_types::sources::SourceConnector;
 use expr::{DummyHumanizer, ExprHumanizer, GlobalId, MirScalarExpr};
-use expr_test_util::generate_explanation;
 use lazy_static::lazy_static;
 use lowertest::*;
 use ore::now::{EpochMillis, NOW_ZERO};
@@ -29,8 +27,6 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::time::{Duration, Instant};
 use uuid::Uuid;
-
-use crate::query_model;
 
 lazy_static! {
     static ref DUMMY_CONFIG: CatalogConfig = CatalogConfig {
@@ -278,7 +274,7 @@ pub enum TestCatalogCommand {
 }
 
 impl TestCatalog {
-    fn execute_commands(&mut self, spec: &str) -> Result<String, String> {
+    pub(crate) fn execute_commands(&mut self, spec: &str) -> Result<String, String> {
         let mut stream_iter = tokenize(spec)?.into_iter();
         loop {
             let command: Option<TestCatalogCommand> = deserialize_optional(
@@ -324,68 +320,4 @@ impl TestCatalog {
         }
         Ok("ok\n".to_string())
     }
-}
-
-#[test]
-fn test_hir_generator() {
-    datadriven::walk("tests/querymodel", |f| {
-        let mut catalog = TestCatalog::default();
-
-        f.run(move |s| -> String {
-            let build_stmt = |stmt, dot_graph, lower| -> String {
-                let scx = &StatementContext::new(None, &catalog);
-                if let sql_parser::ast::Statement::Select(query) = stmt {
-                    let planned_query = match crate::plan::query::plan_root_query(
-                        scx,
-                        query.query,
-                        QueryLifetime::Static,
-                    ) {
-                        Ok(planned_query) => planned_query,
-                        Err(e) => return format!("unable to plan query: {}: {}", s.input, e),
-                    };
-
-                    let model = query_model::Model::from(planned_query.expr);
-
-                    let mut output = String::new();
-
-                    if dot_graph {
-                        output += &match model.as_dot(&s.input) {
-                            Ok(graph) => graph,
-                            Err(e) => return format!("graph generation error: {}", e),
-                        };
-                    }
-
-                    if lower {
-                        output +=
-                            &generate_explanation(&catalog, &model.into(), s.args.get("format"));
-                    }
-
-                    output
-                } else {
-                    format!("invalid query: {}", s.input)
-                }
-            };
-
-            let execute_command = |dot_graph, lower| -> String {
-                let stmts = match sql_parser::parser::parse_statements(&s.input) {
-                    Ok(stmts) => stmts,
-                    Err(e) => return format!("unable to parse SQL: {}: {}", s.input, e),
-                };
-
-                stmts.into_iter().fold("".to_string(), |c, stmt| {
-                    c + &build_stmt(stmt, dot_graph, lower)
-                })
-            };
-
-            match s.directive.as_str() {
-                "build" => execute_command(true, false),
-                "lower" => execute_command(false, true),
-                "cat" => match catalog.execute_commands(&s.input) {
-                    Ok(ok) => ok,
-                    Err(err) => err,
-                },
-                _ => panic!("unknown directive: {}", s.directive),
-            }
-        })
-    });
 }
